migrate getMetadataFor to parent
