+++ modules/@angular/compiler-cli/src/reflector_host.ts
@@ -148,7 +148,7 @@ export class ReflectorHost implements StaticReflectorHost, ImportGenerator {
   private rewriteGenDirPath(filepath: string) {
     var nodeModulesIndex = filepath.indexOf(NODE_MODULES);
     if (nodeModulesIndex !== -1) {
-      // If we are in node_modulse, transplant them into `genDir`.
+      // If we are in node_modules, transplant them into `genDir`.
       return path.join(this.genDir, filepath.substring(nodeModulesIndex));
     } else {
       // pretend that containing file is on top of the `genDir` to normalize the paths.
@@ -231,25 +231,34 @@ export class ReflectorHost implements StaticReflectorHost, ImportGenerator {
     return result;
   }
 
+  /**
+   * @param filePath a canonical file path
+   * @returns Metadata read
+   */
   getMetadataFor(filePath: string): ModuleMetadata {
-    if (!this.context.fileExists(filePath)) {
-      // If the file doesn't exists then we cannot return metadata for the file.
-      // This will occur if the user refernced a declared module for which no file
-      // exists for the module (i.e. jQuery or angularjs).
-      return;
-    }
-    if (DTS.test(filePath)) {
-      const metadataPath = filePath.replace(DTS, '.metadata.json');
-      if (this.context.fileExists(metadataPath)) {
-        const metadata = this.readMetadata(metadataPath);
-        return (Array.isArray(metadata) && metadata.length == 0) ? undefined : metadata;
+    for (const root of this.options.rootDirs || ['', this.options.rootDir || '.']) {
+      const rootedPath = path.join(root, filePath);
+      console.error('try', rootedPath);
+      if (!this.compilerHost.fileExists(rootedPath)) {
+        // If the file doesn't exists then we cannot return metadata for the file.
+        // This will occur if the user referenced a declared module for which no file
+        // exists for the module (i.e. jQuery or angularjs).
+        continue;
       }
-    } else {
-      const sf = this.program.getSourceFile(filePath);
-      if (!sf) {
-        throw new Error(`Source file ${filePath} not present in program.`);
+      if (DTS.test(rootedPath)) {
+        const metadataPath = rootedPath.replace(DTS, '.metadata.json');
+        if (this.context.fileExists(metadataPath)) {
+          const metadata = this.readMetadata(metadataPath);
+          return (Array.isArray(metadata) && metadata.length == 0) ? undefined : metadata;
+        }
+      } else {
+        const sf = this.program.getSourceFile(rootedPath);
+        if (!sf) {
+          throw new Error(`Source file ${rootedPath} not present in program.`);
+        }
+        sf.fileName = this.compilerHost.getCanonicalFileName(sf.fileName);
+        return this.metadataCollector.getMetadata(sf);
       }
-      return this.metadataCollector.getMetadata(sf);
     }
   }
